Make humans and anthropomorphics neutral to each other.
Also implement the actual concept of neutrality in the mob logics: When the relationship between two creatures is too close to 0 (by deault between -0.1 and 0.1), mobs will only take actions when provoked (hit) but not idly. When you are neutral to a creature, their alliance field will now appear yellow in the right-click menu.